pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "Django>=5,<6",
  "Faker>=37.1.0,<38",
  "commitizen>=4.5.0,<5",
  "coveralls>=4.0.1,<5",
  "django-cleanup>=9.0.0,<10",
  "django-cors-headers>=4.7.0,<5",
  "django-debug-toolbar>=5.1.0,<6",
  "django-extensions>=3.2.3,<4",
  "django-structlog>=9.1.0,<10",
  "djangorestframework>=3,<4",
  "environs>=14.1.1,<15",
  "factory-boy>=3.3.3,<4",
  "freezegun>=1.5.1,<2",
  "ipython>=8.35.0",
  "mock>=5.2.0,<6",
  "psycopg-pool>=3.2.6,<4",
  "psycopg[binary]>=3.2.6,<4",
  "pytest-cov>=6.1.1,<7",
  "pytest-django>=4.11.1,<5",
  "pytest-mock>=3.14.0,<4",
  "pytest-xdist>=3.6.1,<4",
  "structlog>=25.2.0,<26"
]
docs = [
  "Sphinx>=7.0.0,<8",
  "myst-parser>=4.0.1",
  "sphinx_rtd_theme"
]
lint = [
  "black>=25.1.0,<26",
  "pre-commit>=3.4.0,<4.0.0",
  "ruff>=0.11.3,<0.12",
  "toml-sort>=0.24.2,<0.25"
]
test = [
  "coverage[toml]>=7.8.0,<8",
  "pytest-cov>=6.1.0,<7",
  "pytest-mock>=3.14.0,<4",
  "pytest-xdist>=3.6.1,<4",
  "pytest>=8.3.5,<9"
]

[project]
authors = [{email = "danghuy1999@gmail.com", name = "Huy Nguyen"}]
classifiers = [
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development"
]
dependencies = [
  "drf-spectacular>=0.28.0,<1"
]
description = "Additional extensions and utilities for drf-spectacular."
license = {file = "LICENSE"}
name = "drf_spectacular_extras"
readme = "README.rst"
requires-python = ">=3.10,<4.0"
version = "0.0.1"

[project.urls]
Documentation = "https://drf-spectacular-extras.readthedocs.io/"
Homepage = "https://github.com/huynguyengl99/drf-spectacular-extras"
Repository = "https://github.com/huynguyengl99/drf-spectacular-extras"

[tool.black]
line-length = 88
preview = true

[tool.commitizen]
gpg_sign = true
name = "cz_conventional_commits"
pre_bump_hooks = [
  "scripts/update_lock.sh"
]
tag_format = "v$version"
update_changelog_on_bump = true
use_shortcuts = true
version_provider = "pep621"
version_scheme = "pep440"

[tool.hatch.build.targets.sdist]
include = ["drf_spectacular_extras"]

[tool.hatch.build.targets.wheel]
include = ["drf_spectacular_extras"]

[tool.tomlsort]
all = true
in_place = true
spaces_before_inline_comment = 2
overrides."lint.isort.section-order".inline_arrays = false

[tool.uv]
default-groups = "all"
